requirements.txt
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonlines==4.0.0
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
requests-mock==1.12.1
six==1.16.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
celery>=5.2.7
redis>=4.3.4
celery[redis]>=5.2.7
django-celery-beat>=2.2.1
pytest>=7.0.1
pytest-django>=4.5.2